Relocate server dir under lib
