+++ .graphqlconfig.yml
@@ -1,12 +1,12 @@
 projects:
   instrumentacs:
-    schemaPath: server/graphql/schema.json
+    schemaPath: lib/graphql/schema.json
     includes:
-      - server/graphql/**/*.ts
+      - lib/graphql/**/*.ts
     excludes:
       - ./amplify/**
     extensions:
       amplify:
         codeGenTarget: typescript
-        generatedFileName: server/api.ts
-        docsFilePath: server/graphql
+        generatedFileName: lib/api.ts
+        docsFilePath: lib/graphql
