+++ pages/_app.tsx
@@ -5,7 +5,7 @@ import 'core-js/features/string/replace-all';
 import Amplify from '@aws-amplify/core';
 import 'katex/dist/katex.min.css';
 import type { AppProps } from 'next/app';
-import config from '../server/aws-exports';
+import config from '../lib/aws-exports';
 import '../styles/globals.css';
 
 Amplify.configure(config);
